requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
astroid==2.12.4
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
dill==0.3.5.1
discord.py==2.0.0
frozenlist==1.3.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
multidict==6.0.2
platformdirs==2.5.2
pycodestyle==2.9.1
pycparser==2.21
pylint==2.15.0
PyNaCl==1.5.0
python-dotenv==0.20.0
rfc3986==1.5.0
sniffio==1.2.0
tekore==4.4.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
wrapt==1.14.1
yarl==1.8.1
youtube-dl==2021.12.17
